requirements.txt
Pillow
folium
streamlit
streamlit_folium 
shapely
geojson
 

